kafka_mini_project/generator/requirements.txt
kafka-python
transactions
